+++ tests_and_analysis/test/euphonic_test/test_force_constants.py
@@ -228,7 +228,12 @@ class TestForceConstantsCreation:
         # again to avoid false positives
         ('NaCl_default', {}),
         ('NaCl_prim', {'summary_name': 'phonopy_nacl.yaml'}),
-        ('CaHgO2', {'summary_name': 'mp-7041-20180417.yaml'})])
+        ('CaHgO2', {'summary_name': 'mp-7041-20180417.yaml'}),
+        ('CaHgO2', {'summary_name': 'phonopy_fullfc.yaml'}),
+        ('CaHgO2', {'summary_name': 'phonopy_nofc.yaml',
+                    'fc_name': 'FULL_FORCE_CONSTANTS'}),
+        ('CaHgO2', {'summary_name': 'phonopy_nofc.yaml',
+                    'fc_name': 'full_force_constants.hdf5'})])
     def create_from_phonopy(self, request):
         material, phonopy_args = request.param
         phonopy_args['path'] = get_fc_dir(material)