+++ CHANGELOG.rst
@@ -6,6 +6,11 @@
   - ``yaml.CSafeLoader`` is now used instead of ``yaml.SafeLoader`` by
     default, so Phonopy ``.yaml`` files should load faster
 
+- Bug fixes:
+
+  - Fix read of Phonopy 'full' force constants from phonopy.yaml and
+    FORCE_CONSTANTS files
+
 `v0.3.2 <https://github.com/pace-neutrons/Euphonic/compare/v0.3.1...v0.3.2>`_
 ----------